100% Scala. Remaining issue:
java.lang.NullPointerException
  at cilly.PEFile.typeDef(PEFile.scala:263)
  at cilly.TableDump.dumpRow(TableDump.scala:227)
  at cilly.TableDump.dumpRow(TableDump.scala:127)
  at cilly.TableDump.dump(TableDump.scala:109)
  at cilly.TableDump.dump(TableDump.scala:91)
  at cilly.TableDump$.main(TableDump.scala:42)

+++ src/main/scala/cilly/emit/FieldBuilder.scala
@@ -19,7 +19,7 @@ import java.io.IOException
  * @author Nikolay Mihaylov
  * @version 1.0
  */
-class FieldBuilder(name: String, declType: Type, attrs: Int, fieldTypeWithMods: PECustomMod)
+class FieldBuilder(name: String, declType: Type, attrs: Short, fieldTypeWithMods: PECustomMod)
   extends FieldInfo(name, declType, attrs, fieldTypeWithMods, null)
   with ICustomAttributeSetter
   with Visitable {
